Move class to its own module
